app/backend/requirements.txt
azure-identity == 1.14.0
Flask == 2.3.3
langchain >= 0.0.157
openai == 0.28.0
azure-search-documents == 11.4.0b3
azure-storage-blob == 12.17.0
azure-cosmos == 4.5.0
opencensus-ext-azure == 1.1.9
tiktoken == 0.4.0
requests-oauthlib == 1.3.1